refactor: 아예 long으로 계산 하도록 변경

+++ src/main/java/com/lhj/myoppingmall/order/entity/Order.java
@@ -118,12 +118,12 @@ public class Order {
     * 전체 주문 가격 계산 메서드
     * */
     private static Long calcTotalPrice(List<OrderItem> orderItems) {
-        int result = 0;
+        long result = 0L;
         if (!orderItems.isEmpty()) {
             for (OrderItem orderItem : orderItems) {
-                result += (int) (orderItem.getOrderPrice() * orderItem.getOrderQuantity());
+                result += orderItem.getOrderPrice() * orderItem.getOrderQuantity();
             }
         }
-        return (long) result;
+        return result;
     }
 }
